Refactor server_model and move create interfaces folder
