+++ server/app/http_api_helpers.py
@@ -20,14 +20,14 @@ import werkzeug.utils
 from werkzeug.exceptions import BadRequest, UnprocessableEntity
 from werkzeug.wrappers import Request, Response
 
-import server.app.server_model as server_model
 from basyx.aas import model
 from basyx.aas.adapter._generic import XML_NS_MAP
 
 from basyx.aas.adapter.xml import xml_serialization, XMLConstructables
 
-from server.app.adapter.xmlization import ServerXMLConstructables, read_server_aas_xml_element
-from server.app.adapter.jsonization import ServerAASToJsonEncoder, ServerStrictAASFromJsonDecoder, ServerStrictStrippedAASFromJsonDecoder
+import server_model
+from .adapter.xmlization import ServerXMLConstructables, read_server_aas_xml_element
+from .adapter.jsonization import ServerAASToJsonEncoder, ServerStrictAASFromJsonDecoder, ServerStrictStrippedAASFromJsonDecoder
 
 from typing import Callable, Dict, List, Optional, Type, TypeVar, Union
