+++ server/app/server_model/__init__.py
@@ -1,0 +1,2 @@
+from .endpoint import *
+from .descriptor import *
